Implemented getAccount and getBalance functionality

+++ src/main/java/com/fidelity/mts/repo/AccountRepo.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import com.fidelity.mts.entity.Account;
 
 @Repository
-public interface AccountRepo extends JpaRepository<Account, Integer>{
+public interface AccountRepo extends JpaRepository<Account, Long>{
 
 }
